Flag to choose alt MicroPython with SoundEffect.
We also need matching stubs so reorg to tie these closer together.
