+++ bin/update-typeshed.sh
@@ -5,5 +5,9 @@
 # Assumes it is run from the root of the editor project and the stubs project is a sibling checkout.
 #
 set -euxo pipefail
-(cd ../micropython-microbit-stubs && python3 scripts/browser-package.py)
-cp ../micropython-microbit-stubs/typeshed.*.json src/language-server/
+cd ../micropython-microbit-stubs
+python3 scripts/browser-package.py
+cd -
+branch=$(git rev-parse --abbrev-ref HEAD)
+mkdir -p "src/micropython/${branch}/"
+cp ../micropython-microbit-stubs/typeshed.*.json "src/micropython/${branch}/"
